% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_blup.R
\name{blup_blup}
\alias{blup_blup}
\title{Compute the unmodified blup of a lmerMod object}
\usage{
blup_blup(model)
}
\arguments{
\item{model}{a lmerMod object.}
}
\description{
Compute the unmodified blup of a lmerMod object
}
