% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_cgr.R
\name{blup_cgr}
\alias{blup_cgr}
\title{Compute the CGR residuals from a lmerMod object}
\usage{
blup_cgr(model)
}
\arguments{
\item{model}{a lmerMod object.}
}
\description{
Compute the CGR residuals from a lmerMod object.
}
