% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_contrblup.R
\name{blup_contrblup}
\alias{blup_contrblup}
\title{Compute the contrasts of the unmodified blup of a lmerMod object}
\usage{
blup_contrblup(model)
}
\arguments{
\item{model}{a lmerMod object.}
}
\description{
Compute the contrasts unmodified blup of a lmerMod object
}
