% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_contrcgr.R
\name{blup_contrcgr}
\alias{blup_contrcgr}
\title{Compute the contrasts of the CGR residuals from a lmerMod object modified}
\usage{
blup_contrcgr(model)
}
\arguments{
\item{model}{a gANOVA object.}
}
\description{
Compute the contrasts of the CGR residuals from a lmerMod object.
}
