% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_lm.R
\name{blup_lm}
\alias{blup_lm}
\title{Compute the blup as fixed effect of a lmerMod object}
\usage{
blup_lm(model)
}
\arguments{
\item{model}{a lmerMod object.}
}
\description{
Compute the blup as fixed effect of a lmerMod object
}
