% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBetweenVar.R
\name{getBetweenVar}
\alias{getBetweenVar}
\title{Get the type of variable}
\usage{
getBetweenVar(formula, data, SU)
}
\arguments{
\item{formula}{a formula of a fixed design}

\item{data}{a dataframe}
}
\description{
get the type of variable of a fixed design given a sampling unit
}
