% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContrlist.R
\name{getContrlist}
\alias{getContrlist}
\title{Extract a list of contrasts of the random design from gANOVA model}
\usage{
getContrlist(model)
}
\arguments{
\item{model}{a lmerModgANOVA model}
}
\description{
Extract a list of contrasts of the random design from gANOVA model
}
