% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpbs.R
\name{getpbs}
\alias{getpbs}
\title{Get one or multiple permutation/bootstrap/shuffle from a pbsmat object}
\usage{
getpbs(x, j)
}
\arguments{
\item{x}{a PBSmat object.}

\item{j}{the indicator of the transformation to get.}
}
\description{
Get one or multiple permutation/bootstrap/shuffle from a pbsmat object
}
