% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZtlm.R
\name{getZtlm}
\alias{getZtlm}
\title{Extract / create the Z matrix of type LM from a model}
\usage{
getZtlm(model, type = "full", ...)
}
\arguments{
\item{model}{a lmerModgANOVA model}

\item{type}{integer indicating the type of matrix "full" for Z0, "reduced" for Z = Z0C' or "contrasts" for C}

\item{...}{othder argument}
}
\description{
Extract / create the Z matrix of type LM from a model
}
