% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4formula2aov.R
\name{lme4formula2aov}
\alias{lme4formula2aov}
\title{Transform a lme4 formula into an aov-like formula}
\usage{
lme4formula2aov(x)
}
\arguments{
\item{x}{a lmerMod or formula object.}
}
\description{
Transform a lme4 formula into a aov-like formula
}
