% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerModperm.R
\name{lmerModperm.list}
\alias{lmerModperm.list}
\title{Compute quasi F from a from gANOVA_lFormula()}
\usage{
\method{lmerModperm}{list}(model, blupstar = "cgr", np = 4000,
  method = "terBraak", assigni = 1, statistic = "quasiF_logp", ...)
}
\description{
Compute quasi F from a from gANOVA_lFormula()
}
