% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerModperm.R
\name{lmerModperm}
\alias{lmerModperm}
\title{Permutation test on lmerMod object}
\usage{
lmerModperm(model, blup_FUN, ...)
}
\arguments{
\item{model}{a lmerMod object}

\item{blup_FUN}{the type of BLUP to permute. Default is CGR.}

\item{...}{other arguments}
}
\description{
permutation test on lmerMod object
}
