% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBS_perm.R
\name{PBS_perm}
\alias{PBS_perm}
\title{Permute vector or matrix using a PBSmat object}
\usage{
PBS_perm(x, PBSmat)
}
\arguments{
\item{x}{vector or matrix}

\item{PBSmat}{PBSmat object}
}
\description{
Permute vector or matrix using a PBSmat object
}
