% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBSmat.R
\name{PBSmat}
\alias{PBSmat}
\title{create permutation, bootstrap or shuffle matrix}
\usage{
PBSmat(n, np = 4000, type = "P")
}
\arguments{
\item{n}{integer, length of vector to transform}

\item{np}{integer, number of tranformations}

\item{type}{character intecating the type of transormation. "P", "S", "B", "PS", "BS" are available.}
}
\description{
create permutation, bootstrap or shuffle matrix
}
