% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranefcontrasts.R
\name{ranefcontrasts}
\alias{ranefcontrasts}
\title{Extract contrast of the random effect for a gANOVA model}
\usage{
ranefcontrasts(model)
}
\arguments{
\item{model}{a gANOVA model}
}
\description{
Extract the contrast (polynomial) of the random effect and with their associated matrix and coding (R contr.poly)
}
