% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitxy.R
\name{refitxy}
\alias{refitxy}
\title{Refit both X and Y of a lme4 model}
\usage{
refitxy(object, newresp = NULL, newx = NULL, rename.response = FALSE, ...)
}
\arguments{
\item{object}{a gANOVA model}

\item{newresp}{a response vector}

\item{newx}{a new X matrix of fixed effect.}

\item{rename.response}{not used.}

\item{...}{other arguments.}
}
\description{
Refit both X and Y of a gANOVA model.
}
