% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerModperm.R
\name{lmerModperm}
\alias{lmerModperm}
\title{Permutation test on lmerMod object}
\usage{
lmerModperm(model, blupstar, np, method, assigni, statistic, ...)
}
\arguments{
\item{model}{a lmerMod object}

\item{blupstar}{a character string indicating the type of random effect to permute. Default is "cgr" and "blup" and "lm"}

\item{np}{an integer indicating the number of permutation/bootstrap/suffle.}

\item{method}{a character string indicating the permutation/bootstrap/suffle method, default is "terBraak".}

\item{assigni}{a integer indicating the effect to test. Default is 1.}

\item{statistic}{a character string indicating the test statistic. default is "Satterthwaite".}

\item{...}{other arguments}
}
\description{
permutation test on lmerMod object
}
